+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The full documentation for rocCV is available at [TBD](TBD)
 
-##  (unreleased) rocCV 0.2.0
+##  rocCV 0.2.0 for ROCm 7.0.0
 
 ### Changed
 
@@ -10,7 +10,8 @@ The full documentation for rocCV is available at [TBD](TBD)
 
 ### Known issues
 
-* Installation on CentOS/RedHat/SLES requires the manual installation of the `DLPack` & `OpenCV` dev packages.
+* Installation on CentOS/RedHat/SLES requires the manual installation of the `DLPack` and `OpenCV` packages.
+* On Ubuntu 22.04 use `pip3 install numpy~=1.23`
 
 ### Upcoming changes